Move fold,flatMap e map in ResponseExtension
